multiple files: move code from evse,cpp to new network.cpp and network.h, introduce debug.h, move include Arduino.h to evse.h
