+++ SmartEVSE-3/platformio.ini
@@ -54,6 +54,7 @@ build_flags =
     -Wall
     -Wextra
     -Wunused-variable
+    -Wno-missing-field-initializers
     -DNDEBUG
     -DMG_ENABLE_LOG=0
     -D MO_DBG_LEVEL=0
@@ -71,8 +72,9 @@ build_flags =
     -Wall
     -Wextra
     -Wunused-variable
+    -Wno-missing-field-initializers
     -DNDEBUG
-    -DVERSION='("v3.6.10+ " __DATE__ " " __TIME__)'
+    -DVERSION='("v3.6.11+ " __DATE__ " " __TIME__)'
 
 [env:debug_telnet]
 lib_deps = ${env.lib_deps}
@@ -87,8 +89,9 @@ build_flags =
     -Wall
     -Wextra
     -Wunused-variable
+    -Wno-missing-field-initializers
     -UNDEBUG
-    -DVERSION='("v3.6.10+D1 " __DATE__ " " __TIME__)'
+    -DVERSION='("v3.6.11+D1 " __DATE__ " " __TIME__)'
 
 
 [env:debug_ser]
@@ -104,6 +107,7 @@ build_flags =
     -Wall
     -Wextra
     -Wunused-variable
+    -Wno-missing-field-initializers
     -DNDEBUG
-    -DVERSION='("v3.6.10+D2 " __DATE__ " " __TIME__)'
+    -DVERSION='("v3.6.11+D2 " __DATE__ " " __TIME__)'
