+++ SmartEVSE-3/src/meter.cpp
@@ -1,4 +1,3 @@
-#include <Arduino.h>
 #include <meter.h>
 #include <modbus.h>
