+++ SmartEVSE-3/src/utils.cpp
@@ -22,7 +22,6 @@
 ; THE SOFTWARE.
  */
 
-#include <Arduino.h>
 #include <stdio.h>
 #include <stdlib.h>
